Delta Calculator
Pushing the delta calculator to main branch. As it's not meant to be accessible by the user, I haven't rendered into the home component, but it is imported there for testing if you want to do something. Please check with each other first before doing so , so you don't stack data in firebase. MAKE SURE TO CHANGE THE INTERVAL PROPERLY IF YOU'RE GOING TO MESS WITH IT.